Finally implemented the mechanism to do formal verification in a nice way.
See tests2_feb2022/test6_formal_verification_timer.py to see how bmc and coverage
tests can be applied to any desired dut module. (e.g. if one module is the
new module to test, and another is a module that demonstrates how to use
the underlying module).

This is great! I had been putting this off for ages.